+++ app/Http/Controllers/ProductController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 use App\Http\Requests\ProductRequest;
+use App\Http\Requests\ProductUpdateRequest;
 use Illuminate\Http\Request;
 use App\Product;
 use App\Category;
@@ -17,7 +18,7 @@ class ProductController extends Controller
      */
     public function index()
     {
-        $products = Product::with('category')->paginate(20);
+        $products = Product::with('category')->withTrashed()->paginate(1);
         return view('products.index', compact('products'));
     }
 
@@ -85,13 +86,14 @@ class ProductController extends Controller
      */
     public function edit($id)
     {
-       
-        $products = Product::find($id);
-        $categories = Category::all();
-        $statuses = ['Active', 'Inactive'];
-        
-        return view('products.create',compact('products', 'statuses', 'categories'));
-      
+        $product = Product::find($id);
+        if($product) {
+            $categories = Category::all();
+            $statuses = ['Active', 'Inactive'];
+            
+            return view('products.create',compact('product', 'statuses', 'categories'));
+        }
+        abort(404);
     }
 
     /**
@@ -101,17 +103,31 @@ class ProductController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(ProductUpdateRequest $request, $id)
     {
         $product = Product::find($id);
-        $product->name = $request->input('name');
-        $product->category_id=$request->input('category');
-        $product->status = $request->input('status');
-        $product->save();
-        \Toastr::success('Point Update successfully', 'Success', []);
-        
-        return redirect()->route('home')
-                        ->with('success','Department update successfully');
+        if($product) {
+            $product->name = $request->input('name');
+            $product->category_id = $request->input('category');
+            $product->status = $request->input('status');
+
+            if($request->has('image')) {
+                $image = $request->file('image');
+                $path = 'products/' . $product->id . '/image/';
+                $filename = Str::random(16) . '.' . $image->getClientOriginalExtension();
+
+                $this->imageUpload($image, $path, $filename);
+
+                $product->image = $filename;
+            }
+
+            $product->update();
+            
+            \Toastr::success('Product Updated successfully', 'Success', []);
+            
+            return redirect()->route('home');
+        }
+        abort(404);
     }
 
     /**
@@ -120,14 +136,21 @@ class ProductController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy($id, Request $request)
     {
-        $product = Product::find($id);
+        $product = Product::withTrashed()->find($id);
 
-        $product->delete();
-        \Toastr::success('Department Deleted successfully', 'Success', []);
+        if($product) {
+            if($request->query('restore') !== null) {
+                $product->restore();
+            } else {
+                $product->delete();
+            }
+            
+            \Toastr::success('Product Deleted successfully', 'Success', []);
 
-        return redirect()->route('home')
-                        ->with('success','Point deleted successfully');
+            return redirect()->route('home');
+        }
+        abort(404);
     }
 }
